reorganized Components: Renderers are now under `Renderers` subdir. Renamed "Render.cpp|h" to "Renderer.cpp|h" to reflect new class name and role. Moved "TextRednerer" files to "Components/Renderers" where they belong. Made "Component/" plural to reflect it holds a collection of componenets.
