+++ src/CLGEngine/CMakeLists.txt
@@ -2,5 +2,5 @@ add_library(CLGEngine Input.cpp Time.cpp Game.cpp EntityManager.cpp Entity.cpp T
 
 add_subdirectory(Graphics)
 add_subdirectory(CORE)
-add_subdirectory(Component)
-target_link_libraries(CLGEngine PUBLIC CORE Graphics Component)
+add_subdirectory(Components)
+target_link_libraries(CLGEngine PUBLIC CORE Graphics Components)
