--- src/CLGEngine/Component/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_library(Component Component.h Collider.cpp Render.cpp BlockRenderer.cpp)
