+++ src/CLGEngine/Components/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(Components Component.h Collider.cpp)
+
+add_subdirectory(Renderers)
+target_link_libraries(Components PUBLIC Renderers)
