+++ src/CLGEngine/Components/Renderers/BlockRenderer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Render.h"
+#include "Renderer.h"
 
 namespace CLGEngine {
     class BlockRenderer : public Renderer
