+++ src/CLGEngine/Components/Renderers/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_library(Renderers Renderer.cpp TextRenderer.cpp BlockRenderer.cpp)
