+++ src/CLGEngine/Components/Renderers/TextRenderer.cpp
