+++ src/Game/Entities/Character.h
@@ -3,9 +3,9 @@
 #include "../GameManager.h"
 #include "../../CLGEngine/TileMap.h"
 #include "../../CLGEngine/CORE/Vector2.h"
-#include "../../CLGEngine/Component/Collider.h"
+#include "../../CLGEngine/Components/Collider.h"
 #include "../../CLGEngine/Entity.h"
-#include "../../CLGEngine/Component/BlockRenderer.h"
+#include "../../CLGEngine/Components/Renderers/BlockRenderer.h"
 
 
 class Character : public CLGEngine::Entity {
