+++ src/Game/Entities/ScreenText.h
@@ -2,7 +2,7 @@
 
 #include "../../CLGEngine/CORE/Vector2.h"
 #include "../../CLGEngine/Entity.h"
-#include "../../CLGEngine/Graphics/TextRenderer.h"
+#include "../../CLGEngine/Components/Renderers/TextRenderer.h"
 #include <string>
 
 class ScreenText
